requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
python-dotenv==1.0.1

sqlalchemy==2.0.31
psycopg[binary]==3.2.9
redis==5.0.7

stripe==9.10.0
sendgrid==6.11.0
httpx==0.27.0

python-dateutil==2.9.0.post0
pytz==2024.1
rapidfuzz==3.9.3
reportlab==4.2.5

# Optional helper
twilio==9.2.3
